argc: ast: move symbol collector to seperate file
